Suppress warning from MSVC about converting from size_t to int

+++ include/nanogui/vector.h
@@ -340,7 +340,7 @@ public:
      * The alpha component of the color, will be divided by ``255.0``.
      */
     Color(size_t intensity, int alpha)
-        : Color(Vector3i(intensity), alpha) { }
+        : Color(Vector3i(static_cast<int>(intensity)), alpha) { }
 
     /**
      * Explicit constructor: creates the Color ``(r, g, b, a)``.
